+++ CHANGES.md
@@ -1,6 +1,7 @@
 Changes coming in 0.4.2
 
 * Add `metadata-query` macro to make metadata query / results easier to work with [JDBC-107](http://dev.clojure.org/jira/browse/JDBC-107).
+* `prepare-statement` `:return-keys` may now be a vector of (auto-generated) column names to return, in addition to just being truthy or falsey. [JDBC-104](http://dev.clojure.org/jira/browse/JDBC-104).
 
 Changes in 0.4.0 / 0.4.1